fix-clang-warnings

+++ dlls/rpcrt4/ndr_marshall.c
@@ -711,7 +711,7 @@ static inline ULONG safe_multiply(ULONG a, ULONG b)
 
 static inline void safe_buffer_increment(MIDL_STUB_MESSAGE *pStubMsg, ULONG size)
 {
-    if ((pStubMsg->Buffer + size < pStubMsg->Buffer) || /* integer overflow of pStubMsg->Buffer */
+    if ((pStubMsg->Buffer + (LONG)size < pStubMsg->Buffer) || /* integer overflow of pStubMsg->Buffer */
         (pStubMsg->Buffer + size > (unsigned char *)pStubMsg->RpcMsg->Buffer + pStubMsg->BufferLength))
         RpcRaiseException(RPC_X_BAD_STUB_DATA);
     pStubMsg->Buffer += size;
@@ -719,7 +719,7 @@ static inline void safe_buffer_increment(MIDL_STUB_MESSAGE *pStubMsg, ULONG size
 
 static inline void safe_buffer_length_increment(MIDL_STUB_MESSAGE *pStubMsg, ULONG size)
 {
-    if (pStubMsg->BufferLength + size < pStubMsg->BufferLength) /* integer overflow of pStubMsg->BufferSize */
+    if (pStubMsg->BufferLength + (LONG)size < pStubMsg->BufferLength) /* integer overflow of pStubMsg->BufferSize */
     {
         ERR("buffer length overflow - BufferLength = %lu, size = %lu\n",
             pStubMsg->BufferLength, size);
@@ -732,7 +732,7 @@ static inline void safe_buffer_length_increment(MIDL_STUB_MESSAGE *pStubMsg, ULO
  * to do so */
 static inline void safe_copy_from_buffer(MIDL_STUB_MESSAGE *pStubMsg, void *p, ULONG size)
 {
-    if ((pStubMsg->Buffer + size < pStubMsg->Buffer) || /* integer overflow of pStubMsg->Buffer */
+    if ((pStubMsg->Buffer + (LONG)size < pStubMsg->Buffer) || /* integer overflow of pStubMsg->Buffer */
         (pStubMsg->Buffer + size > pStubMsg->BufferEnd))
     {
         ERR("buffer overflow - Buffer = %p, BufferEnd = %p, size = %lu\n",
@@ -748,7 +748,7 @@ static inline void safe_copy_from_buffer(MIDL_STUB_MESSAGE *pStubMsg, void *p, U
 /* copies data to the buffer, checking that there is enough space to do so */
 static inline void safe_copy_to_buffer(MIDL_STUB_MESSAGE *pStubMsg, const void *p, ULONG size)
 {
-    if ((pStubMsg->Buffer + size < pStubMsg->Buffer) || /* integer overflow of pStubMsg->Buffer */
+    if ((pStubMsg->Buffer + (LONG)size < pStubMsg->Buffer) || /* integer overflow of pStubMsg->Buffer */
         (pStubMsg->Buffer + size > (unsigned char *)pStubMsg->RpcMsg->Buffer + pStubMsg->BufferLength))
     {
         ERR("buffer overflow - Buffer = %p, BufferEnd = %p, size = %lu\n",
@@ -767,7 +767,7 @@ static void validate_string_data(MIDL_STUB_MESSAGE *pStubMsg, ULONG bufsize, ULO
     ULONG i;
 
     /* verify the buffer is safe to access */
-    if ((pStubMsg->Buffer + bufsize < pStubMsg->Buffer) ||
+    if ((pStubMsg->Buffer + (LONG)bufsize < pStubMsg->Buffer) ||
         (pStubMsg->Buffer + bufsize > pStubMsg->BufferEnd))
     {
         ERR("bufsize 0x%lx exceeded buffer end %p of buffer %p\n", bufsize,
